08_05 comp
In this exercise we are going to create a program that can be used to create and modify a to-do list. The final product will work in the following manner.

Sample output
Command: add
Task: go to the store
Command: add
Task: vacuum clean
Command: list
1: go to the store
2: vacuum clean
Command: completed
Which task was completed? 2
Task go to the store tehty
Command: list
1: go to the store
Command: add
Task: program
Command: list
1: go to the store
2: program
Command: stop