08_10 task "I owe you"
Create a class called IOU which has the following methods:

constructor public IOU() creates a new IOU
public void setSum(String toWhom, double amount) saves the amount owed and the person owed to to the IOU.
public double howMuchDoIOweTo(String toWhom) returns the amount owed to the person whose name is given as a parameter. If the person cannot be found, it returns 0.