task 08_11 Hash for date
Let's expand the SimpleDate class from the previous exercise to also have its own hashCode method.

Create a method public int hashCode() for the SimpleDate class, which calculates a hash for the the SimpleDate object. Implement the calculation of the hash in way that there are as few similar hashes as possible between the years 1900 and 2100.