09_03 Warehousing (7 parts)
Product warehouse, step 1
The class Warehouse handles the functions related to the amount of a product. Now we want a product name for the product and a way to handle the name. Let's write ProductWarehouse as a subclass of Warehouse! First, we'll just create a private object variable for the product name, a constructor, and a getter for the name field:

public ProductWarehouse(String productName, double capacity) - Creates an empty product warehouse. The name of the product and the capacity of the warehouse are provided as parameters.
public String getName() - Returns the name of the product.

Product warehouse, step 2
As we can see from the previous example, the toString() inherited by the ProductWarehouse object doesn't (obviously!) know anything about the product name. Something must be done! Let's also add a setter for the product name while we're at it:

public void setName(String newName) - sets a new name for the product.
public String toString() - Returns the state of the object represented as a string like this Juice: balance = 64.5, space left 123.5

Change History, step 1
Public constructors and methods of the ChangeHistory class:

public ChangeHistory() creates an empty ChangeHistory object.
public void add(double status) adds provided status as the latest amount to remember in the change history.
public void clear() empties the history.
public String toString() returns the string representation of the change history. The string representation provided by the ArrayList class is sufficient.

Change History, step 2
Build on the ChangeHistory class by adding analysis methods:

public double maxValue() returns the largest value in the change history. If the history is empty, the method should return zero.
public double minValue() returns the smallest value in the change history. If the history is empty, the method should return zero.
public double average() returns the average of the values in the change history. If the history is empty, the method should return zero.
The methods should not modify the order of the encapsulated list.

Product warehouse with history, step 1
Public constructors and methods:

public ProductWarehouseWithHistory(String productName, double capacity, double initialBalance) creates a product warehouse. The product name, capacity, and initial balance are provided as parameters. Set the initial balance as the initial balance of the warehouse, as well as the first value of the change history.
public String history() returns the product history like this [0.0, 119.2, 21.2]. Use the string representation of the ChangeHistory object as is.

Product warehouse with history, step 2
It's time to make history! The first version didn't know anything but the initial state of the history. Expand the class with the following methods

public void addToWarehouse(double amount) works just like the method in the Warehouse class, but we also record the changed state to the history. NB: the value recorded in the history should be the warehouse's balance after adding, not the amount added!
public double takeFromWarehouse(double amount) works just like the method in the Warehouse class, but we also record the changed state to the history. NB: the value recorded in the history should be the warehouse's balance after removing, not the amount removed!

Product warehouse with history, step 3
Expand the class with the method

public void printAnalysis(), which prints history related information for the product in the way presented in the example.