09_04 DifferentKindsOfBoxes (3 parts)
Implement the equals and hashCode methods for the Item-class.

Implement the class BoxWithMaxWeight, that inherits the Box class. BoxWithMaxWeight has a constructor public BoxWithMaxWeight(int capacity), that defines the max weight allowed for that box.

Next, implement the class OneItemBox, that inherits the Box class. OneItemBox has the constructor public OneItemBox(), and it has the capacity of exactly one item. If there is already an item in the box, it must not be switched. The weight of the item added to the box is irrelevant.