09_05 TacoBoxes (2 parts)
the method int tacosRemaining() return the number of tacos remaining in the box.
the method void eat() reduces the number of tacos remaining by one. The number of tacos remaining can't become negative.

Implement the class TripleTacoBox, that implements the TacoBox interface. TripleTacobox has a constructor with no parameters. TripleTacobox has an object variable tacos which is initialized at 3 when the constructor is called.

Implement the class CustomTacoBox, that implements the TacoBox interface. CustomTacoBox has a constructor with one parameter defining the initial number of tacos in the box(int tacos).