09_11 Herds (2 points)
part 1
Create a class called Organism that implements the interface Movable. An organism should know its own location (as x, y coordinates). The API for the class Organism is to be as follows:
public Organism(int x, int y)
The class constructor that receives the x and y coordinates of the initial position as its parameters.
public String toString()
Creates and returns a string representation of the organism. That representation should remind the following: "x: 3; y: 6". Notice that a semicolon is used to separate the coordinates.
public void move(int dx, int dy)
Moves the object by the values it receives as parameters. The dx variable contains the change to coordinate x, and the dy variable ontains the change to the coordinate y. For example, if the value of dx is 5, the value of the object variable x should be incremented by five.

part 2
Implementing the Herd
Create a class called Herd that implements the interface Movable. A herd consists of multiple objects that implement the Movable interface. They must be stored in e.g. a list data structure.
The Herd class must have the following API.
public String toString()
Returns a string representation of the positions of the members of the herd, each on its own line.
public void addToHerd(Movable movable)
Adds an object that implements the Movable interface to the herd.
public void move(int dx, int dy)
Moves the herd with by the amount specified by the parameters. Notice that here you have to move each member of the herd.