10_13 Literacy Comparison
This exercise uses open data about literacy levels, provided by UNESCO. The data includes statistics of estimated or reported levels of literacy for various countries for the past two years. File literacy.csv, included with the exercise template, includes the literacy estimates for women and men over 15 years of age. Each line in the file literacy.csv is as follows: "theme, age group, gender, country, year, literacy percent. Below are the first five lines as an example.

Create a program that first reads the file literacy.csv and then prints the contents from the lowest to the highest ranked on the literacy rate. The output must be exactly as in the following example. The example shows the first five of the expected rows.