10_14 Literature 3 parts
part 1 : Reading and printing the books
Implement the reading and printing the books first, the ordering of them doesn't matter yet.

part 2: Ordering books based on their age recommendation
Expand your program so, that the books are sorted based on their age recommendations when they are printed. If two (or more) books share the same age recommendations the order between them does not matter.

part 3 : Ordering books based on their age recommendation and name
Expand your program, so that it sorts the books with the same age recommendation based on their name alphabetically.