10_15 Regular Expressions
part 1 Day of week
Use regular expressions to create the method public boolean isDayOfWeek(String string), which returns true if the parameter string is an abbreviation of a day of the week (mon, tue, wed, thu, fri, sat, sun)

part 2 Vowel check
NB. For simplicity's sake, in this exercises the letters that are considered vowels are: a, e, i, o, and u.
Create the method public boolean allVowels(String string) that uses a regular expression to check whether all the characters in the parameter string are vowels.

part 3 Time of day
Create the method public boolean timeOfDay(String string). It should use a regular expression to check whether the parameter string expresses a time of day in the form hh:mm:ss (hours, minutes, and seconds each always take up two spaces).
NB. In this exercise we use the 24-hour clock. So the acceptable values are between 00:00:00 and 23:59:59.