11_10 Flight Control
In this exercise, you will implement a flight control application. It is used to control the airplanes and their flight routes. The system always knows the identifier and the capacity of an airplance. The flight information consists of the used airplane, the departure airport id (e.g. HEL), and the destination airport ID (e.g. BAL).

There can be multiple airplanes and flights. The same airplane can be used to make several flights.

The application should operate in two parts: first, the user enters information about airplanes and flights in the airport asset control, after which the program offers the flight information service for the user. There are three operations in this latter flight control — printing the airplanes, printing the flights, and printing the information of a single airplane. In addition, the user may exit the program by choosing the option x. If the user enters an invalid command, the program asks for a command again.

The program should start when the main method of the Main class inside the package FlightControl is executed