11_11 Validating Parameters
Validating a person:
The constructor of the class Person should ensure that the name given as the parameter is not null, empty, or over 40 characters in length. The age should between 0 and 120. If some of these conditions are not met, the constructor should throw an IllegalArgumentException.

Validating the calculator:
The methods of the Calculator class should be as follows: The method factorial should only work if the parameter is a non-negative number (0 or greater). The method binomialCoefficient should only work when the parameters are non-negative and the subset size does not exceed the set size. If these methods receive invalid parameters when they are called, they should throw an IllegalArgumentException