Part 11-13.SaveableDictionary
Completed part 1
For the dictionary, implement a constructor which takes no parameters, and the following methods:

public void add(String words, String translation) adds a word to the dictionary. Every word has just one translation, and if the same word is added for the second time, nothing happens.
public String translate(String word) returns the translation for the given word. If the word is not in the dictionary, returns null.