Part 11-13.SaveableDictionary
Completed part 2 and 3

Part 2 Deleting words
Add the dictionary method public void delete(String word) which deletes the given word and its translation from the dictionary.

It might be worth reviewing the material concerning deleting objects from an ArrayList from previous weeks.

Part 3 Reading from file
Make a constructor public SaveableDictionary(String file) and method public boolean load(), which loads the dictionary from the file given to the constructor as a parameter. If the program is unable to open the file or read from it, the method returns false, otherwise it returns true.

NB: the constructor only tells the dictionary the name of the file to load the dictionary from. The constructor does not read the file. Only the method load reads the file.
In the dictionary file, one line contains a word and its translation separated by ":".