sum these for me
In the class Program implement a class method public static int sum(int[] array, int fromWhere, int toWhere, int smallest, int largest). The method must calculate the sum of the elements in the array between the lower and the upper limits. Only numbers smaller or equal to the int largest and larger or equal to the int smallest are added to the sum.

The method must also check that the lower and the upper limit are valid indexes in the array. If the parameter fromWhere is smaller than 0, the lower limit becomes 0 instead. Accordingly, if the parameter toWhere is larger than the size of the array, the upper limit becomes the last index of the array instead.